Fixed model + controller folder structures + namespace issue in decorator
